just reset break after sleep
When the computer is put to sleep, it generally means the user is away
from the computer and doing something else, or the user is moving the
laptop to another location. In either case, we assume the user had a
break and just stop to annoy users with breaks shortly after sleep
awake.