fix pause immediately after break end if idle
The original implementation is buggy. If the user is busy after break
for a long while, and suddenly idle for 1s, then it will be recognized
as an idle pause, which is not want I intended.

With the current implementation, the watcher starts before break
windows. I hope thing will go well.