dirty setting detection and explicit sync
Use controllers to manage settings state:
- is dirty or not
- load from preferences
- save to preferences
- trigger events for syncing labels

With these we have:
- Dirty detection. Ask users to save, discard or cancel when switching
tab or closing the preference window
- Explict sync. Previously, if the current preference value is the same
as the default one, the labels won't be synced. Therefore, it relies on
the UI file to provide a correct initial state. However, having to
translate so many strings like `5 min`, `20 sec` is tedious. With this
commit, we don't need them at all.
- Save preferences as needed. Some users may need to modify the config
file directly if the settings presented in the UI is not satisfactory,
e.g. the user need a extra long duration for breaks. If the user open
the preference window and edit some other settings, that setting will
also change because the Qt widget can't represent the value in the file.
With this commit, the user can have customized values in the preference
file while being able to use the UI to configure other settings.

See also:
- #28
- #29#issuecomment-2703596807