remove reading mode for stability
I have received some bug reports about reading mode that make the app
stop working. For example, the timer is not ticking or the app freezes.
I have no idea why this happens and there are different problems with
different devices. I have to remove this feature to make the app stable.

Reopens #11