enable immersive break experience
Specifically, this commit
- Add a clock and time when break ends to the full-screen window
- Users can choose which UI elements are shown on windows
- It's possible to leave the window open until user activity

By doing this, the user can remove progress bar and countdown from the
window and, instead of focusing on the countdown, the user is encouraged
to take other activities and come back after the break ends. Also, by
not closing the window as soon as the break ends, we are not encouraging
user to come back to work immediately.

Closes #60