merge gui and app subdirs
The split between gui and app is meaningless and it makes the app subdir
contain only a few files. Besides, If we plan to add preview feature
when configuring the appearance of break windows, the `break-windows.h`
must be contained in the gui subdir, making the app subdir even smaller.
Therefore, the best approach is to merge them and get cleaner dirs.