just prevent any new instances to start up
Starting more than one instances is not useful for ordinary users, and
the Qt `QLockFile` is actually very robust to crashes. And allowing
other instances to start can cause problems:
- This is confusing and not useful for ordinary users.
- If user starts another instance, and then closes the first one, when
the user starts the third one, there will not be dialogs.

The reason why I allowed new instances to start:
- I thought `QLockFile` is not robust and can create false positives
where the lock file exists but the program is not running. If this
happens, I have to provide a way for the user to start another instance.
- When developing Sane Break, I need one for tracking my time and
another one to debug.

The first reason is not valid after I reading through the documentation
more carefully, and the second one is achieved by using a different lock
file name for debug builds.

Closes #82