do not make window transparent for inputs on X11
X11 does not support changing transparency for inputs after the window
is created. To make the experience as consistent as other platforms, we
decide to not make it transparent to inputs at all, so that the buttons
on the full-screen window is clickable and won't actually click on apps
behind it.

Closes #92