smaller clock size with countdown & remove seconds
The clock is too "noisy" and attracts unnecessary attention if it shows
seconds. Even though it might be useful when the countdown is hidden, in
reality, it does not. That's because the use case of hiding a countdown
and showing a clock is when the break is long. Therefore, the seconds
makes almost no difference in a ~10 min break. Thus, we'd better remove
it.

Besides, we should also consider the use case where the user want both
the countdown and the clock. In this case, we should make the font size
of the clock smaller.

Closes #87